% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createK}
\alias{createK}
\title{Create a Kernel matrix from a design matrix G}
\usage{
createK(G)
}
\arguments{
\item{G}{A numeric matrix.}
}
\value{
A normalized kernel matrix K.
}
\description{
Create a Kernel matrix from a design matrix G
}
