% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMNMF_main.R
\name{group_onehot}
\alias{group_onehot}
\title{Compute one-hot matrix for given data frame and variable(s)}
\usage{
group_onehot(x, ivar)
}
\arguments{
\item{x}{Input data frame.}

\item{ivar}{Variable(s) for one-hot computation.}
}
\value{
A one-hot encoded matrix.
}
\description{
Compute one-hot matrix for given data frame and variable(s)
}
